rename test to github_runner_on_gitea_test.go
* merge portforward_test.go into test
